remove usize

+++ glit-core/src/org.rs
@@ -66,7 +66,7 @@ impl OrgFactory {
             pages_urls,
             all_branches: self.all_branches,
             repositories_data: DashMap::<_, _, RandomState>::with_capacity_and_hasher(
-                repo_count as usize,
+                repo_count,
                 RandomState::new(),
             ),
         }
